dashboard/requirements.txt
requests
pytest
pytest-cov
pandas
python-dotenv
streamlit
altair
psycopg2-binary
sqlalchemy
matplotlib
seaborn
